Include Magic Hash in base_url for resilience
The server side shutterbug endpoint expects all iframes to
have a "magic hash" in them, else it fails to decode them
properly. Chrome's innerHTML encodes the payload using
encodeURIComponent, but Firefox does not. Thus, Chrome's
payload always works, but Firefox's only works when the
URL includes a hash. Adding a hash to Chrome's URL does
not affect its output.

By inspecting the URL to see if it already has a hash, and
adding if it doesn't, we ensure that all sent URLs will be
rendered properly.

See
https://github.com/concord-consortium/shutterbug/commit/a203755475b32144b775d3de8d5306bc6a2bb9e1
for details.